requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.8
docopt==0.6.2
einops==0.6.1
future==0.18.2
gast==0.5.4
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.36.0
h5py==3.9.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.0
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numpy==1.19.1
nvidia-dali-cuda110==1.29.0
nystrom-attention==0.0.11
oauthlib==3.1.0
olefile==0.46
opencv-python==4.3.0.36
openslide-python==1.3.0
pandas==1.1.0
Pillow==7.2.0
protobuf==3.15.3
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.2
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1605205327372/work
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.2.26
torch==1.7.1
torchaudio==0.7.2
# torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.48.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.3
Werkzeug==1.0.1
